added CopyButton use client component
