--- app/t.css
@@ -1,55 +1,0 @@
-/* latin-ext */
-@font-face {
-    font-family: 'Vollkorn';
-    font-style: italic;
-    font-weight: 400;
-    font-display: swap;
-    src: url(https://fonts.gstatic.com/s/vollkorn/v22/0yb7GDoxxrvAnPhYGxksWEEeqTo29TfO-Q.woff2) format('woff2');
-
-}
-
-/* latin-ext */
-@font-face {
-    font-family: 'Vollkorn';
-    font-style: italic;
-    font-weight: 500;
-    font-display: swap;
-    src: url(https://fonts.gstatic.com/s/vollkorn/v22/0yb7GDoxxrvAnPhYGxksWEEeqTo29TfO-Q.woff2) format('woff2');
-}
-
-/* latin */
-@font-face {
-    font-family: 'Vollkorn';
-    font-style: italic;
-    font-weight: 500;
-    font-display: swap;
-    src: url(https://fonts.gstatic.com/s/vollkorn/v22/0yb7GDoxxrvAnPhYGxksWE8eqTo29Tc.woff2) format('woff2');
-}
-
-
-/* latin-ext */
-@font-face {
-    font-family: 'Vollkorn';
-    font-style: normal;
-    font-weight: 400;
-    font-display: swap;
-    src: url(https://fonts.gstatic.com/s/vollkorn/v22/0yb9GDoxxrvAnPhYGxknaE0Urhg0xTY.woff2) format('woff2');
-}
-
-/* latin-ext */
-@font-face {
-    font-family: 'Vollkorn';
-    font-style: normal;
-    font-weight: 500;
-    font-display: swap;
-    src: url(https://fonts.gstatic.com/s/vollkorn/v22/0yb9GDoxxrvAnPhYGxknaE0Urhg0xTY.woff2) format('woff2');
-}
-
-/* latin-ext */
-@font-face {
-    font-family: 'Vollkorn';
-    font-style: normal;
-    font-weight: 700;
-    font-display: swap;
-    src: url(https://fonts.gstatic.com/s/vollkorn/v22/0yb9GDoxxrvAnPhYGxknaE0Urhg0xTY.woff2) format('woff2');
-}
