+++ components/atomic/buttons/CopyButton/index.js
@@ -1,0 +1,2 @@
+export * from './ButtonWrapper';
+export { default } from './ButtonWrapper';
