+++ middleware/hooks/useClipboard.js
@@ -1,3 +1,4 @@
+'use client'
 import { useState, useCallback } from 'react';
 
 function useClipboard() {
